refactor(schemas): single-source approval action sets in the schema layer
BASH_APPROVAL_ACTIONS and TOOL_EDIT_APPROVAL_ACTIONS were declared twice —
once in the runtime modules (@tools/approval/{bashApproval,toolEditApproval})
and again, byte-identically, in the progressView inbound IPC schema — kept in
sync by hand. Adding an action to one side without the other would make the
inbound schema reject an action the backend emits.

Hoist both const+type into @shared/schemas/prompts (which already owns
PLAN_APPROVAL_ACTIONS / USER_QUESTION_ACTIONS and has no @tools/@agent imports).
The IPC schema imports them directly; the runtime modules re-export them so
existing importers of @tools/approval/* are unchanged. Direction is @tools ->
@shared/schemas (never the reverse), so no runtime code leaks into the schema
layer / renderer bundle and no cycle forms.
