+++ src/tools/approval/bashApproval.ts
@@ -3,6 +3,10 @@ import { z } from 'zod';
 import type { AgentRuntimeHost } from '@agent/runtime/AgentRuntimeHost';
 import { tryUseRunContext } from '@agent/runtime/RunContext';
 import { StreamTabIdSchema, type StreamTabId } from '@shared/schemas';
+import {
+  BASH_APPROVAL_ACTIONS,
+  type BashApprovalAction,
+} from '@shared/schemas/prompts';
 import { requireRuntimeHost } from '@tools/contextHelpers';
 import { type ToolResult } from '@tools/result';
 import { getConfig } from '@utils/config/configUtils';
@@ -25,9 +29,9 @@ export type BashApprovalResult = z.infer<typeof BashApprovalResultSchema>;
 
 export const BASH_APPROVAL_CONFIG_KEY = 'texra.toolUse.requireBashApproval';
 
-export const BASH_APPROVAL_ACTIONS = ['approve', 'reject'] as const;
-
-export type BashApprovalAction = (typeof BASH_APPROVAL_ACTIONS)[number];
+// Defined in the schema layer (see @shared/schemas/prompts); re-exported here
+// so existing importers of `@tools/approval/bashApproval` keep working.
+export { BASH_APPROVAL_ACTIONS, type BashApprovalAction };
 
 const DEFAULT_BASH_REJECTION_INSTRUCTION =
   'Do not retry this rejected command or another approval-gated shell command for the same check. ' +
