+++ src/tools/approval/toolEditApproval.ts
@@ -11,6 +11,10 @@ import { tryUseRunContext } from '@agent/runtime/RunContext';
 import { isLatexFile } from '@common/files/fileTypeUtils';
 import { StreamTabIdSchema, type StreamTabId } from '@shared/schemas';
 import {
+  TOOL_EDIT_APPROVAL_ACTIONS,
+  type ToolEditApprovalAction,
+} from '@shared/schemas/prompts';
+import {
   LineChangesSchema,
   type LineChanges,
   type ToolResult,
@@ -50,16 +54,9 @@ const TOOL_EDIT_APPROVAL_CONFIG_KEY = 'texra.toolUse.requireEditApproval';
 
 export const REVEAL_TIMEOUT_MS = 1500;
 
-export const TOOL_EDIT_APPROVAL_ACTIONS = [
-  'approve',
-  'reject',
-  'openDiff',
-  'showLatexdiff',
-  'previewProposed',
-] as const;
-
-export type ToolEditApprovalAction =
-  (typeof TOOL_EDIT_APPROVAL_ACTIONS)[number];
+// Defined in the schema layer (see @shared/schemas/prompts); re-exported here
+// so existing importers of `@tools/approval/toolEditApproval` keep working.
+export { TOOL_EDIT_APPROVAL_ACTIONS, type ToolEditApprovalAction };
 
 export const toolEditApprovalController =
   createStreamApprovalController<ToolEditApprovalResult>({
